+++ modules/datastore/src/Service/Factory/Import.php
@@ -3,61 +3,22 @@
 namespace Drupal\datastore\Service\Factory;
 
 use Drupal\datastore\Storage\DatabaseTableFactory;
-use Drupal\datastore\Service\ImportService;
 use Drupal\common\Storage\JobStoreFactory;
 
 /**
  * Create an importer object for a given resource.
+ *
+ * @deprecated
+ * @see \Drupal\datastore\Service\Factory\ImportServiceFactory
  */
-class Import implements ImportFactoryInterface {
-
-  /**
-   * Job store factory.
-   *
-   * @var \Drupal\common\Storage\JobStoreFactory
-   */
-  private $jobStoreFactory;
-
-  /**
-   * Database table factory.
-   *
-   * @var \Drupal\datastore\Storage\DatabaseTableFactory
-   */
-  private $databaseTableFactory;
-
-  /**
-   * Import services.
-   *
-   * @var \Drupal\datastore\Service\ImportService[]
-   */
-  private $services = [];
+class Import extends ImportServiceFactory {
 
   /**
    * Constructor.
    */
   public function __construct(JobStoreFactory $jobStoreFactory, DatabaseTableFactory $databaseTableFactory) {
-    $this->jobStoreFactory = $jobStoreFactory;
-    $this->databaseTableFactory = $databaseTableFactory;
-  }
-
-  /**
-   * Inherited.
-   *
-   * @inheritdoc
-   */
-  public function getInstance(string $identifier, array $config = []) {
-
-    if (!isset($config['resource'])) {
-      throw new \Exception("config['resource'] is required");
-    }
-
-    $resource = $config['resource'];
-
-    if (!isset($this->services[$identifier])) {
-      $this->services[$identifier] = new ImportService($resource, $this->jobStoreFactory, $this->databaseTableFactory);
-    }
-
-    return $this->services[$identifier];
+    parent::__construct($jobStoreFactory, $databaseTableFactory);
+    @trigger_error(__NAMESPACE__ . '\Import is deprecated. Use \Drupal\datastore\Service\Factory\ImportServiceFactory instead.', E_USER_DEPRECATED);
   }
 
 }
