+++ modules/datastore/tests/src/Unit/EventSubscriber/DatastoreSubscriberTest.php
@@ -18,7 +18,7 @@ use MockChain\Chain;
 use MockChain\Options;
 use PHPUnit\Framework\TestCase;
 use Symfony\Component\DependencyInjection\Container;
-use Drupal\datastore\Service\Factory\Import as ImportServiceFactory;
+use Drupal\datastore\Service\Factory\ImportServiceFactory as ImportServiceFactory;
 use Drupal\datastore\Service\ImportService;
 use Drupal\common\Storage\JobStore;
 use Drupal\metastore\MetastoreItemInterface;
