+++ modules/datastore/tests/src/Unit/ServiceTest.php
@@ -9,7 +9,7 @@ use Drupal\common\DataResource;
 use Drupal\common\Storage\JobStore;
 use Drupal\common\Storage\JobStoreFactory;
 use Drupal\datastore\DatastoreService;
-use Drupal\datastore\Service\Factory\Import as ImportServiceFactory;
+use Drupal\datastore\Service\Factory\ImportServiceFactory as ImportServiceFactory;
 use Drupal\datastore\Service\ImportService;
 use Drupal\datastore\Service\Info\ImportInfoList;
 use Drupal\datastore\Service\ResourceLocalizer;
